Refactor: TimeDealCreateRequest에 아이템 세부 정보를 내부 record로 통합 및 디렉토리로 이동(domain.timeDeal -> domain.item)
