--- src/main/java/com/jelly/zzirit/domain/item/dto/timeDeal/request/TimeDealCreateItemDetail.java
@@ -1,7 +1,0 @@
-package com.jelly.zzirit.domain.item.dto.timeDeal.request;
-
-public record TimeDealCreateItemDetail(
-	Long itemId,
-	int quantity
-) {
-}
